backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "cryptography>=43.0.1",
    "fastapi[standard]>=0.115.0",
    "groq>=0.11.0",
    "pymysql>=1.1.1",
    "python-dotenv>=1.0.1",
    "python-multipart>=0.0.12",
    "ruff>=0.6.9",
    "sqlalchemy>=2.0.35",
]
